backend/requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3,<4.4
psycopg2-binary>=2.9,<2.10
djangorestframework-simplejwt>=5.3,<5.4
django-filter>=24.0
Pillow>=10.0
reportlab>=4.0
